+++ .vitepress/theme/index.ts
@@ -1,5 +1,16 @@
-import defaultTheme from 'vitepress/theme'
 import './custom.css'
 
+import defaultTheme from 'vitepress/theme'
+import PackageVersion from '../vitepress/components/PackageVersion.vue'
+import type { Theme } from 'vitepress'
 
-export default defaultTheme
+
+const theme: Theme = {
+  ...defaultTheme,
+  enhanceApp({ app }) {
+    app.component('PackageVersion', PackageVersion)
+  }
+}
+
+
+export default theme